added entities, controllers, services and repos for pending items

+++ src/main/java/com/garofalo/medicare_simplilearn_backend/repository/ProductManufacturerRepository.java
@@ -7,5 +7,5 @@ import org.springframework.web.bind.annotation.CrossOrigin;
 
 @CrossOrigin("http://localhost:4200")
 @RepositoryRestResource(collectionResourceRel = "productManufacturer", path="product-manufacturer")
-public interface ProductManufacturerRepository extends JpaRepository<ProductManufacturer, Integer> {
+public interface ProductManufacturerRepository extends JpaRepository<ProductManufacturer, Long> {
 }
