+++ src/main/java/com/garofalo/medicare_simplilearn_backend/repository/ProductRepository.java
@@ -9,12 +9,12 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import org.springframework.web.bind.annotation.CrossOrigin;
 
 @CrossOrigin("http://localhost:4200")
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 
     //find items by category. path will be 'api/products/search/findByCategoryId?id=XX'
     //Similar to @Query. Will provide a list of all the elements based on the id
     //Data Rest will automatically expose the endpoint
-    Page<Product> findByCategoryId(@Param("id") int id, Pageable pageable);
+    Page<Product> findByCategoryId(@Param("id") Long id, Pageable pageable);
 
     //find items by name. path will be api/products/search/findByNameContaining?name=XXX
     //Data Rest will automatically expose the endpoint
